+++ R/utils.R
@@ -32,3 +32,25 @@ is_workdir_clean <- function(repo) {
   current_status <- status(repo)
   all(vapply(current_status, length, integer(1)) == 0)
 }
+
+#' Check if a vector contains valid email
+#' @param email A vector with email addresses.
+#' @return A logical vector.
+#' @export
+#' @importFrom assertthat assert_that
+validate_email <- function(email) {
+  assert_that(is.character(email))
+  # expression taken from https://emailregex.com/
+  grepl(
+    paste0(
+      "(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"",
+      "(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|",
+      "\\[\x01-\x09\x0b\x0c\x0e-\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*",
+      "[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:25[0-5]|",
+      "2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9]", #nolint
+      "[0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a",
+      "\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\\])"
+    ),
+    email
+  )
+}